refactor(persistence): alinha DDL ao diagrama ER
O diagrama ER é a fonte da verdade do schema; onde o DDL divergia e a
correção não quebrava a aplicação, o código foi ajustado:

- F000_PREFERENCES: TXT_KEY -> TXT_PREFERENCE VARCHAR(20) (PK passa a ser
  COD_PERSON + TXT_PREFERENCE), TXT_VALUE VARCHAR(255) -> VARCHAR(60), e
  ganha TMS_CREATE_AT/TMS_UPDATED_AT — acompanhado por
  PreferencesJDBCRepository e ClosingJDBCRepository, que agora gravam os
  dois timestamps.
- F004_TAG: ganha FLG_ACTIVE e TMS_UPDATED_AT. FLG_ACTIVE existe só no
  schema — a tag não tem ciclo de vida no domínio, então fica 'Y' fixo
  até haver regra de negócio que o mova.
- F002_BALANCE: ganha os timestamps, NUM_PERIOD vira NUMERIC(6) e a chave
  de negócio vira UNIQUE (COD_ACCOUNT, NUM_PERIOD). ID segue como PK
  física porque findById/deleteById assumem PK simples.
- SYS_TRANSACTION_NATURE: TXT_DESCRIPTION VARCHAR(50) -> VARCHAR(10).
- F000_COST_CENTER: TXT_DESCRIPTION VARCHAR(255) -> VARCHAR(50).
- F006_TRANSACTION: NUM_SIGNAL INT -> NUMERIC(1).

Sem migração one-shot no projeto: um banco de desenvolvimento anterior a
esta mudança não sobe. Apague database.mv.db/database.trace.db antes de
rodar.

+++ br-application/src/main/java/br/cdb/feature/f006/F006Module.java
@@ -43,7 +43,7 @@ public class F006Module implements Lifecycle {
                     ID CHAR(36) PRIMARY KEY,
                     COD_PERSON CHAR(36),
                     TXT_DESCRIPTION VARCHAR(255) NOT NULL,
-                    NUM_SIGNAL INT NOT NULL,
+                    NUM_SIGNAL NUMERIC(1) NOT NULL,
                     DEC_AMOUNT DECIMAL(19, 2) NOT NULL,
                     TMS_PURCHASE TIMESTAMP NOT NULL,
                     COD_ACCOUNT CHAR(36) NOT NULL,
